+++ packages/shopify-app-kit/tsup.config.ts
@@ -4,6 +4,8 @@ export default defineConfig({
   entry: {
     index: "src/index.ts",
     core: "src/core.ts",
+    error: "src/error.ts",
+    http: "src/http.ts",
     feedback: "src/feedback.ts",
     "save-flow": "src/save-flow.ts",
     navigation: "src/navigation.ts",